requirements.txt
python-dotenv
flask
langchain
openai
slack-sdk
slack-bolt
google-generativeai